Added overlapping_words which returns the number of overlapping words in two strings.
Slightly annoyed at having to write this but M$ mixerGetID and
mixerOpen fn's do not work for wave identifiers so there is no way to
tie the mixer to wave in / wave out devices bar matching manufacturer
id and finding the most overlap in words between the mixer name and
wave name.  Grrr! Thwack!