- Change socket_error to support formatted messages. - Add socket_error messages for gethostbyname failures.
Got a complaint of freebsd ports RAT crashing immediately on start-up.
After being given an account (and root privelages :) on a machine in
Russia, it was tracked down to the machine having a different name to
the name in DNS for the IP address (and no entry in /etc/hosts).

Additional socket error messages should make this immediately obvious
in future. [oh]