Added code in mbus_recv function to ignore duplicate reliable messages:
The sequence number of the last reliable message is stored in an associative array with the source as the key.  If the same sequence number arrives from the same source then the packet is ignored.

The problem occurs when the ACK is delayed by too much MBus traffic and the sender resends the same reliable message.  This is common when one of the processes is delayed on startup and the other waiting process will flood the MBus channel with mbus.waiting messages.  When the two processes finally sync and send reliable messages there is a backlog of mbus.waiting messages to process and the ACK for the reliable messages is lost within the backlog.  Thus causing the sender to resend the message.

Kristian.