Got grabber partially working with yuv file - the colour information seems to come after the luminosity information so at the moment the image is just grey- scale. Will need to tweak things to get YUV colour information.
Added basic frame sequence playback but it shows the that frames contain
more information that we're reading so you see a whole frame followed by a 
half a frame and half color info.

The fix from the previous versio was to re-instate the allocref(); call in the
setsize() function. Also to reinstate the height_ and width_ variables - 
although they're declared in module.h - the Grabber is not a Module.