This was done by Piers O'Hanlon.
Make sure that do not discard the very first two data packets. 

This is because Vic-cc will not clock packets out if the sender doesn't hear 
from the receiver.  I.e., if the receiver discard the very first data packet, 
then the sender is not likely send the second data packet.