-- modifed suspend_grabbing() base funtion to avoid compile-time error, e.g.:
error: cannot allocate an object of abstract type ‘JpegEncoder’
note:   because the following virtual functions are pure within ‘JpegEncoder’:

this error occurred because suspend_grabbing() is declared as pure virtual
function earlier, hence no way to allocate an object of abstract JpegEncoder,
for example.

it is now replaced with an empty function in module.h.