(This changes deal with the packet re-ordering.)
Motivation: there are good number of packet re-ordering, especially when the
part of links are consisted of a wireless channel.

Packet re-ordering is particularly bad for RTT measurement, and for cwnd
calculation.

Upon reordering detection, we skip sampling RTT (but still update RTT using
previously sampled RTT). Also, insert the re-ordered packet sequence number into
the received AckVec using previously received AckVec.

** Future work:
when there are more than 3 packet re-ordering, cwnd is not increased - we may
need to consider inflating cwnd intentionally when there are 3 or more packet
re-ordering detection.
(This is because the packets are actually not lost, but only re-ordered. 
However, in the cwnd calculation, cwnd has been reduced because the 
re-ordered packets are out of the DUPACK range.)