this commit reflects Revision 4701 from the mpeg4 branch.
copied comments follows:
Removed MacOS X 10.3 workaround code originally from Aqua AG vic which I
initially thought was a Quicktime Pthread related issue and wrote so in the
comments.

It was actually to do with the framework initialization not being called.
Framework initialization is not required and shouldn't be used with MacOS X
10.3.9 and later.