This commit reflects Revision 4699 from the mpeg4 branch.
copied comments follows:
Fixed bug when trying to send in unicast and hopefully on Win7 too. In my
connect/disconnect() fix I also 'cleaned' up the sockaddr declarations (sin) -
This was an error as the sin structure was being used by ::sendto() in dosend()
- and relying on it being setup in ssock(). I renamed the sin member variable to
  sin_ (and anotheri(local_preset) so the member variables are '_' postfixed -
good idea in general), and added local declarations of sin as and where needed.