(this commit re-writes the packet re-ordering policy done by Revision 4730)
when packet re-ordering occurred (or ack re-ordering) and it is beyond 3
DUPACKS, then we need to revert to the previous state (reverting to the earlier
average loss interval and re-mark the timestamp).

in short, in Revision 4730, we trigger a packet out when this happens.
with this commit, instead of clocking a packet out, we revert ALI and timestamp
to the previous stage if:
    -- this re-ordered seqno is beyond 3 DUPACKS
    -- and that seqno actually changed ALI earlier