It is critically important for TFWC (or any kind of ack-clock based CC mechanisms) to receive Ack as soon as it arrives to the data sender, so that the data sender can update CC information (cwnd or send rate) in a timely fashion.
We have discovered that Ack can be arrived to the sender in a bulky manner -
i.e., multimple Acks arrived in a very short time scale. 

Although TFWC had a mechanism to pull out the arrived-ack before sending every
data packet, it wasn't sufficient when MANY Acks are arrived already.

Therefore, we have created a while loop to pull out the arrived XR as soon as
possible as long as it is there. So, these Acks are not the normal Ack clock,
hence we have changed some variable names accordingly (recv_by_ch => ack_clock).

Also, there was a typo in tfwc_sndr.h when returning the number of bytes that
this Ack is giving.