** SO_TIMESTAMP is now working fine **    (tested over FreeBSD and Linux)
There are some trick to get it working right.

First, for FreeBSD/Mac OSX, we had to use SCM_TIMESTAMP instead of SO_TIMESTAMP
inside "recvmsg()" method.

Second, for Linux, we had to enable SO_TIMESTAMP when opening socket even in the
sender - normally, we only need to enable SO_TIMESTAMP in the receiver's socket.


Because we can use socket timestamp when measuring RTT, TFWC and TFRC use this
socket timestamp when computing RTT/SRTT.